Update and rename head.py to ww_head.py
Added some more structure, making it easier to distinguish what commands can be used by whom.
